move gc::register_mark_object to Ruby::gc_register_mark_object
same for gc::register_address and gc::unregister_address

This is fallout from removing (/moving to the right place)
Ruby::get_unchecked(). Some methods rely on taking `Value` to
prove they are on a Ruby thread, but turns out this didn't hold
for `T: Mark`, and that became obvious when trying to push
Ruby::get_unchecked() to the right place.